WIP - Modal built, needs toggle logic cleaned up, styling -- more cleanup
