+++ src/config/routes.js
@@ -5,10 +5,7 @@ import NewPost from '../components/NewPost'
 import EditForm from '../components/EditForm'
 import ViewPost from '../components/ViewPost'
 import GardenProfile from '../components/GardenProfile'
-import ViewPostBuy from '../components/ViewPostBuy'
-import ViewPostTrade from '../components/ViewPostTrade'
 import OtherProfiles from '../components/OtherProfiles'
-import ViewPostVolunteer from '../components/ViewPostVolunteer'
 import Volunteer from '../components/Volunteer'
 
 
@@ -20,12 +17,7 @@ export default(
         <Route exact path='/newpost' component={ NewPost } />
         <Route exact path='/editform' component={ EditForm } />
         <Route exact path='/post/:id' component={ ViewPost } />
-        {/* <Route exact path='/viewpost/:id' component={ ViewPost } /> */}
-        {/* <Route exact path='/viewpostbuy/:id' component={ ViewPostBuy } /> */}
-        {/* <Route exact path='/viewposttrade/:id' component={ ViewPostTrade } /> */}
-        {/* <Route exact path='/viewposttrade' component={ ViewPostTrade } /> */}
         <Route exact path='/gardenprofile/:id'  component={GardenProfile} />
-        {/* <Route exact path='/viewpostvolunteer/:id'  component={ViewPostVolunteer}/> */}
         <Route exact path='/previewprofile/:id'  component={OtherProfiles} />
         <Route exact path='/volunteer' component={Volunteer} />
     </Switch>
